feat: better static analysis templates
Ensures that all validation result implementations MUST declare the type they wrap.
Discovered that the reason I was having issues previously was due to the fact that the _named constructors_ needed to define a template, accept the type as a `$value`, and then return `self<TemplateName>`.

Once this was in place, all of it validated properly.