refactor: add Options interface and implementation for RuleSet
Adding options allows simplifying the `RuleSet` dramatically.
The constructor will take an Options instance, and use it to setup it's internal state, allowing it to be idempotent.
The only extension points become the constructor (and the named constructor `createWithRules(Rule ...$rules)`) and the `createValidResultSet()` method.
If users want to change the mechanics of how validation works, they can implement the RuleSetValidator interface.