refactor: refactor `Rule` to make missing, default, return ValidationResult
- Adds `Rule::missing(): ValidationResult`
- Refactors `Rule::default()` to return `ValidationResult`

This change simplifies the logic of a `RuleSet`, as it no longer needs to worry about composing logic for generating a missing value, and it ensures that it ALWAYS has a default value for required values when generating a default valid result set.

Additionally, it makes it easier to provide non-Result ValidationResult implementations for both values, ensuring type safety when needed.

This change allows removing:

- `CreateMissingValueResult`
- `MissingValueResultFactory`
- `RequiredRuleWithNoDefaultValueException`

Required a refactor of the `BooleanRule` and `CallbackRule` to rewrite the `default()` method implementations, and add `missing()` implementations.

Additionally, required removing the `MissingValueResultFactory` from ruleset options.